muhteme

+++ libft_h/ft_putnbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: vparlak <vparlak@student.42kocaeli.com.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/29 04:44:49 by vparlak           #+#    #+#             */
-/*   Updated: 2023/03/12 18:35:35 by vparlak          ###   ########.fr       */
+/*   Updated: 2023/03/12 22:08:46 by vparlak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putnbr_fd(int n, int fd)
+int	ft_putnbr_fd(long n, int fd)
 {
 	long		nbr;
 	static int	count = 0;
